Move parameter helper functions to respective test files and update assertion usage in ArrayParameterTest
